Generates a random sample from a Burr XII distribution

This function is designed to generate random data following a Burr XII distribution with parameters shape1 and shape2.

Usage

rburrxii(n, shape1, shape2)

Arguments

n

number of observations.

shape1, shape2

shape parameters. Must be strictly positive.

Value

A vector containing data distributed according to the Burr XII distribution with a sample size specified by n, and with shape parameters specified by shape1 and shape2.

Details

The Burr XII distribution with parameters shape1 = \(\gamma\) and shape2 = \(\alpha\) has the probability density function: $$f_{X}(x) = \alpha\gamma\dfrac{x^{\alpha-1}}{(1+x^\alpha)^{\gamma+1}}$$ for \(x>0\), \(\gamma>0\) and \(\alpha>0\).

Examples

rburrxii(10, 2, 3)
#>  [1] 0.3503327 1.1336357 0.8352208 0.4469438 0.1549515 0.7172295 0.7435481
#>  [8] 0.5714268 0.9777956 1.0312370
rburrxii(100, 4, 16)
#>   [1] 0.9762248 0.8284666 0.7436532 0.8667026 0.8833020 0.8350937 0.8835332
#>   [8] 0.7739713 0.8806745 1.0268868 0.8598933 0.9326820 0.9434765 0.8351815
#>  [15] 1.0328558 0.9447044 0.7633504 0.9064912 0.9359108 0.9345585 0.7392979
#>  [22] 0.8437575 0.8623861 0.9251186 0.8974418 0.8889186 0.9379050 1.0059720
#>  [29] 0.8302494 0.8413406 0.9330093 0.9009695 0.9260475 0.9293875 0.7951917
#>  [36] 0.9496010 0.9504525 1.0511440 1.0218584 0.8807681 0.8942318 0.8655835
#>  [43] 0.8283781 0.9067296 0.9000460 0.9524956 0.8376565 0.9392279 0.7751892
#>  [50] 0.8738072 0.9629460 0.8561150 0.9134489 0.8699754 0.9180301 0.8338119
#>  [57] 1.0055008 0.9086376 0.9090085 0.8572535 0.8915954 0.8772941 0.7342766
#>  [64] 0.8950995 0.8809329 0.7188155 0.8783763 0.9116811 0.9711886 0.8799160
#>  [71] 0.9060890 0.9187966 0.8530849 0.8599297 0.8976298 0.9920658 0.8829942
#>  [78] 0.8402785 0.9314720 0.7698254 1.0774309 0.8192418 0.9044462 0.9682368
#>  [85] 0.9401608 0.8479826 0.9094347 0.9653240 0.7341015 0.8957117 0.9583362
#>  [92] 0.9602793 0.8836044 0.8417730 0.8863411 0.9309444 0.9025258 0.9294011
#>  [99] 0.9032558 0.9655133