Generates a random sample from a Pareto I distribution

This function is designed to generate random data following a Pareto I distribution with parameters scale and shape.

Usage

rparetoi(n, scale, shape)

Arguments

n

number of observations.

scale

scale parameter. Must be strictly positive.

shape

shape parameter. Must be strictly positive.

Value

A vector containing data distributed according to the Pareto I distribution with a sample size specified by n, and with scale and shape parameters specified by scale and shape, respectively.

Details

The Pareto I distribution with parameters scale = \(\theta\) and shape = \(\alpha\) has the probability density function: $$f_{X}(x) = \dfrac{\alpha\theta^{\alpha}}{x^{\alpha+1}}$$ for \(x\geq \theta\), \(\theta>0\) and \(\alpha>0\).

Examples

rparetoi(10, 5, 2)
#>  [1]  5.644307  5.040495  5.356263  8.923779  8.355893  6.096376  6.388672
#>  [8]  9.245010  9.093245 18.337886
rparetoi(100, 10, 6)
#>   [1] 11.07821 11.63041 10.30512 11.49662 14.63148 11.62331 13.45803 11.61824
#>   [9] 12.86051 10.88194 13.71708 12.53199 10.65811 10.19933 10.17922 13.07767
#>  [17] 10.82917 10.09064 20.49546 11.67032 10.27212 11.37659 10.22659 17.33762
#>  [25] 10.08044 10.29773 15.83155 18.00844 12.29331 14.34193 18.73501 10.05938
#>  [33] 11.01053 11.88102 24.05019 10.64999 13.81063 11.38309 14.11251 11.28402
#>  [41] 14.04028 13.82089 10.73253 10.00017 10.39015 16.22670 10.56604 14.25724
#>  [49] 17.18061 10.00671 11.43578 10.42110 11.89364 12.42108 14.09715 10.83510
#>  [57] 14.50767 13.62759 12.43680 10.06550 10.70044 12.59311 15.02721 10.38830
#>  [65] 11.12229 10.84449 10.86690 10.41910 10.01207 10.78262 11.32539 13.63186
#>  [73] 10.24661 10.81907 10.52250 11.21661 10.03292 10.15320 25.06018 16.07905
#>  [81] 10.22383 10.20606 10.79105 11.11978 12.91166 11.18907 10.18083 16.14137
#>  [89] 12.26431 11.23311 11.34071 10.51093 12.30380 17.79080 11.97613 13.03619
#>  [97] 10.12752 14.33470 13.39267 12.40640