Generates a random sample from a Pareto I distribution

This function is designed to generate random data following a Pareto I distribution with parameters scale and shape.

Usage

rparetoi(n, scale, shape)

Arguments

n

number of observations.

scale

scale parameter. Must be strictly positive.

shape

shape parameter. Must be strictly positive.

Value

A vector containing data distributed according to the Pareto I distribution with a sample size specified by n, and with scale and shape parameters specified by scale and shape, respectively.

Details

The Pareto I distribution with parameters scale = \(\theta\) and shape = \(\alpha\) has the probability density function: $$f_{X}(x) = \dfrac{\alpha\theta^{\alpha}}{x^{\alpha+1}}$$ for \(x\geq \theta\), \(\theta>0\) and \(\alpha>0\).

Examples

rparetoi(10, 5, 2)
#>  [1]  9.265317 14.097430  5.028948 14.957031 82.725680  7.072422  6.244962
#>  [8] 10.538889  7.756603  8.264478
rparetoi(100, 10, 6)
#>   [1] 13.85561 11.49655 10.49815 15.19650 14.00282 10.48775 10.89741 12.77210
#>   [9] 10.21435 10.36748 10.30422 11.98872 13.82173 15.45247 11.43356 11.54218
#>  [17] 12.13891 10.47889 10.07636 14.93400 10.12292 26.17626 11.70848 10.32073
#>  [25] 14.92261 10.06382 11.61893 10.46109 14.83797 13.29442 12.22096 10.42280
#>  [33] 12.42328 10.41632 11.06872 10.69771 11.50302 10.49630 11.09535 15.16114
#>  [41] 18.23992 13.29699 14.75076 11.56186 12.80359 24.28333 12.29825 10.41157
#>  [49] 10.59180 12.37647 13.98994 10.46447 10.00751 16.14407 11.00850 12.60429
#>  [57] 12.01628 10.91297 10.74782 12.50187 11.99513 10.14957 13.81463 10.13441
#>  [65] 13.76271 10.18916 11.16875 10.48468 12.13323 10.30227 16.63595 10.66873
#>  [73] 10.77648 14.39744 13.40967 10.17839 14.83150 12.80098 10.83917 37.71365
#>  [81] 10.74244 16.33265 10.41414 10.05451 10.26516 13.78685 10.40762 10.40137
#>  [89] 10.06743 16.20480 10.47937 12.88150 10.58305 14.14890 10.60178 19.82429
#>  [97] 11.60132 12.67782 13.51738 12.71023